// web/src/proto/workspace/workspace.ts
export interface ListWorkspacesRequest {
    /**
     * @generated from protobuf field: int32 organization_id = 1
     */
    organizationId: number;
    /**
     * @generated from protobuf field: wws.common.PaginationRequest pagination = 2
     */
    pagination?: PaginationRequest;
}
// ...
export interface GetWorkspaceRequest {
    /**
     * @generated from protobuf field: int32 id = 1
     */
    id: number;
}
/**
 * WorkspaceResponse
 *
 * @generated from protobuf message wws.workspace.WorkspaceResponse
 */
export interface WorkspaceResponse {
    /**
     * @generated from protobuf field: wws.workspace.Workspace workspace = 1
     */
    workspace?: Workspace;
}
/**
 * CreateWorkspaceRequest
 *
 * @generated from protobuf message wws.workspace.CreateWorkspaceRequest
 */
export interface CreateWorkspaceRequest {
    /**
     * @generated from protobuf field: string name = 1
     */
    name: string;
    /**
     * @generated from protobuf field: int32 organization_id = 2
     */
    organizationId: number;
    /**
     * @generated from protobuf field: int32 cpu = 3
     */
    cpu: number;
    /**
     * @generated from protobuf field: int32 memory = 4
     */
    memory: number;
    /**
     * @generated from protobuf field: int32 storage = 5
     */
    storage: number;
    /**
     * @generated from protobuf field: repeated string languages = 6
     */
    languages: string[];
    /**
     * @generated from protobuf field: string region = 7
     */
    region: string;
}
/**
 * UpdateWorkspaceRequest
 *
 * @generated from protobuf message wws.workspace.UpdateWorkspaceRequest
 */
export interface UpdateWorkspaceRequest {
    /**
     * @generated from protobuf field: int32 id = 1
     */
    id: number;
    /**
     * @generated from protobuf field: string name = 2
     */
    name: string;
    /**
     * @generated from protobuf field: int32 cpu = 3
     */
    cpu: number;
    /**
     * @generated from protobuf field: int32 memory = 4
     */
    memory: number;
    /**
     * @generated from protobuf field: int32 storage = 5
     */
    storage: number;
    /**
     * @generated from protobuf field: repeated string languages = 6
     */
    languages: string[];
    /**
     * @generated from protobuf field: string region = 7
     */
    region: string;
}
/**
 * DeleteWorkspaceRequest
 *
 * @generated from protobuf message wws.workspace.DeleteWorkspaceRequest
 */
export interface DeleteWorkspaceRequest {
    /**
     * @generated from protobuf field: int32 id = 1
     */
    id: number;
}
/**
 * StartWorkspaceRequest
 *
 * @generated from protobuf message wws.workspace.StartWorkspaceRequest
 */
export interface StartWorkspaceRequest {
    /**
     * @generated from protobuf field: int32 id = 1
     */
    id: number;
}
/**
 * StopWorkspaceRequest
 *
 * @generated from protobuf message wws.workspace.StopWorkspaceRequest
 */
export interface StopWorkspaceRequest {
    /**
     * @generated from protobuf field: int32 id = 1
     */
    id: number;
}
/**
 * RestartWorkspaceRequest
 *
 * @generated from protobuf message wws.workspace.RestartWorkspaceRequest
 */
export interface RestartWorkspaceRequest {
    /**
     * @generated from protobuf field: int32 id = 1
     */
    id: number;
}
// ...
class ListWorkspacesRequest$Type extends MessageType<ListWorkspacesRequest> {
    constructor() {
        super("wws.workspace.ListWorkspacesRequest", [
            { no: 1, name: "organization_id", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 2, name: "pagination", kind: "message", T: () => PaginationRequest }
        ]);
    }
    create(value?: PartialMessage<ListWorkspacesRequest>): ListWorkspacesRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.organizationId = 0;
        if (value !== undefined)
            reflectionMergePartial<ListWorkspacesRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ListWorkspacesRequest): ListWorkspacesRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* int32 organization_id */ 1:
                    message.organizationId = reader.int32();
                    break;
                case /* wws.common.PaginationRequest pagination */ 2:
                    message.pagination = PaginationRequest.internalBinaryRead(reader, reader.uint32(), options, message.pagination);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: ListWorkspacesRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* int32 organization_id = 1; */
        if (message.organizationId !== 0)
            writer.tag(1, WireType.Varint).int32(message.organizationId);
        /* wws.common.PaginationRequest pagination = 2; */
        if (message.pagination)
            PaginationRequest.internalBinaryWrite(message.pagination, writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message wws.workspace.ListWorkspacesRequest
 */
export const ListWorkspacesRequest = new ListWorkspacesRequest$Type();
// ...
class GetWorkspaceRequest$Type extends MessageType<GetWorkspaceRequest> {
    constructor() {
        super("wws.workspace.GetWorkspaceRequest", [
            { no: 1, name: "id", kind: "scalar", T: 5 /*ScalarType.INT32*/ }
        ]);
    }
    create(value?: PartialMessage<GetWorkspaceRequest>): GetWorkspaceRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.id = 0;
        if (value !== undefined)
            reflectionMergePartial<GetWorkspaceRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: GetWorkspaceRequest): GetWorkspaceRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* int32 id */ 1:
                    message.id = reader.int32();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: GetWorkspaceRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* int32 id = 1; */
        if (message.id !== 0)
            writer.tag(1, WireType.Varint).int32(message.id);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message wws.workspace.GetWorkspaceRequest
 */
export const GetWorkspaceRequest = new GetWorkspaceRequest$Type();
// ...
class CreateWorkspaceRequest$Type extends MessageType<CreateWorkspaceRequest> {
    constructor() {
        super("wws.workspace.CreateWorkspaceRequest", [
            { no: 1, name: "name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "organization_id", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 3, name: "cpu", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 4, name: "memory", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 5, name: "storage", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 6, name: "languages", kind: "scalar", repeat: 2 /*RepeatType.UNPACKED*/, T: 9 /*ScalarType.STRING*/ },
            { no: 7, name: "region", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<CreateWorkspaceRequest>): CreateWorkspaceRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.name = "";
        message.organizationId = 0;
        message.cpu = 0;
        message.memory = 0;
        message.storage = 0;
        message.languages = [];
        message.region = "";
        if (value !== undefined)
            reflectionMergePartial<CreateWorkspaceRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: CreateWorkspaceRequest): CreateWorkspaceRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string name */ 1:
                    message.name = reader.string();
                    break;
                case /* int32 organization_id */ 2:
                    message.organizationId = reader.int32();
                    break;
                case /* int32 cpu */ 3:
                    message.cpu = reader.int32();
                    break;
                case /* int32 memory */ 4:
                    message.memory = reader.int32();
                    break;
                case /* int32 storage */ 5:
                    message.storage = reader.int32();
                    break;
                case /* repeated string languages */ 6:
                    message.languages.push(reader.string());
                    break;
                case /* string region */ 7:
                    message.region = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: CreateWorkspaceRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string name = 1; */
        if (message.name !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.name);
        /* int32 organization_id = 2; */
        if (message.organizationId !== 0)
            writer.tag(2, WireType.Varint).int32(message.organizationId);
        /* int32 cpu = 3; */
        if (message.cpu !== 0)
            writer.tag(3, WireType.Varint).int32(message.cpu);
        /* int32 memory = 4; */
        if (message.memory !== 0)
            writer.tag(4, WireType.Varint).int32(message.memory);
        /* int32 storage = 5; */
        if (message.storage !== 0)
            writer.tag(5, WireType.Varint).int32(message.storage);
        /* repeated string languages = 6; */
        for (let i = 0; i < message.languages.length; i++)
            writer.tag(6, WireType.LengthDelimited).string(message.languages[i]);
        /* string region = 7; */
        if (message.region !== "")
            writer.tag(7, WireType.LengthDelimited).string(message.region);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message wws.workspace.CreateWorkspaceRequest
 */
export const CreateWorkspaceRequest = new CreateWorkspaceRequest$Type();
// ...
export const UpdateWorkspaceRequest = new UpdateWorkspaceRequest$Type();
// ...
export const DeleteWorkspaceRequest = new DeleteWorkspaceRequest$Type();
// ...
export const StartWorkspaceRequest = new StartWorkspaceRequest$Type();
// ...
export const StopWorkspaceRequest = new StopWorkspaceRequest$Type();
// ...
export const RestartWorkspaceRequest = new RestartWorkspaceRequest$Type();